% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addscenario.R
\name{addscenario}
\alias{addscenario}
\title{Title}
\usage{
addscenario(p, oldMPA, MPA_coverage = 0.1, replicates = 10, dist = NA,
  name = "Status_quo", Habitats = FALSE, Breeding = FALSE, cell_size)
}
\arguments{
\item{n}{}
}
\description{
Title
}

