% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcnetvalue.R
\name{calcnetvalue}
\alias{calcnetvalue}
\title{Title}
\usage{
calcnetvalue(catch_summary, Status_quo_profitability, fish_operating_cost_ratio)
}
\arguments{
\item{fish_operating_cost_ratio}{}
}
\description{
Title
}

