% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal.R
\name{dispersal}
\alias{dispersal}
\title{Adult or larval dispersal}
\usage{
dispersal(fish, cm, ages = 5:50)
}
\arguments{
\item{ages}{}
}
\description{
Adult or larval dispersal
}
\examples{

cells=5000
cm <- matrix(1/cells,nrow=cells,ncol=cells)
fish <- initpop(initial_abun=250*10^6,cells=cells,maxage=50,rate=0.7)
fish <- dispersal(fish,cm,ages=5:50)
}

