% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatequota.R
\name{estimatequota}
\alias{estimatequota}
\title{Title}
\usage{
estimatequota(fish, maxage, biomass_est_bank, y, tot_time, FMSY, FMSY_buffer)
}
\arguments{
\item{biomass_est_bank}{}
}
\description{
Title
}
\examples{

}

