% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish2eggs.R
\name{fish2eggs}
\alias{fish2eggs}
\title{Title}
\usage{
fish2eggs(fish, fecundity = 0.5 * 10^6, age_mat_steepness = 2.5,
  age_mat_sigmoid = 4, l_to_w_int = 1.1e-05, l_to_w_power = 2.91,
  Linf_mean = 112.03, Linf_SD = 10.46/1.96, k_mean = 0.13,
  k_SD = 0.021/196, t0 = 0.18)
}
\arguments{
\item{l_to_w_power}{}

\item{t0}{}
}
\description{
Title
}
\examples{



}

