% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish2weight.R
\name{fish2weight}
\alias{fish2weight}
\title{Calculate weight from fish age}
\usage{
fish2weight(fish, ages, l_to_w_int = 1.1e-05, l_to_w_power = 2.91,
  Linf_mean = 112.03, Linf_SD = 10.46/1.96, k_mean = 0.13,
  k_SD = 0.021/196, t0 = 0.18)
}
\arguments{
\item{fish}{a matrix of fish numbers by site (rows) and by age (columns). Can easily be generated by \code{initpop}}

\item{ages}{The ages of the fish of the fish to be converted to weight (year)}

\item{l_to_w_int}{The intercept in the Length-weight relationship}

\item{l_to_w_power}{The power in the Length-weight relationship}

\item{Linf_mean}{Von Bertalanffy growth model parameter - mean asymptotic length (cm)}

\item{Linf_SD}{Von Bertalanffy growth model parameter - standard deviation asymptotic length (cm)}

\item{k_mean}{Von Bertalanffy growth model parameters - mean growth coefficient (1/year)}

\item{k_SD}{Von Bertalanffy growth model parameters - standard deviation growth coefficient (1/year)}

\item{t0}{Von Bertalanffy growth model parameters - x intercept (year)}
}
\description{
Wrapper function for \code{\link{age2length}} and \code{\link{length2weight}} to estimate weight in each cell from \code{fish}.
}
\examples{
# create a fish matrix
fish <- initpop(initial_abun=250*10^6,cells=length(BESTMPA_domain),maxage=50,rate=0.7)
# calculate weight
fish2weight(fish,ages=0:20,l_to_w_int=0.000011,l_to_w_power=2.91,Linf_mean=112.03,Linf_SD=10.46/1.96,k_mean=0.13,k_SD=0.021/196,t0=0.18)
}

