% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatempasizes.R
\name{generatempasizes}
\alias{generatempasizes}
\title{Title}
\usage{
generatempasizes(cell_size, cells = 1, p, MPA_coverage, included,
  MPAs_AREA = MPAs_mar_REP_M_AREA)
}
\arguments{
\item{MPAs_AREA}{Vector of MPA size frequency in m^2. Defaults to MPA szie frequency obtained from WDPA database downloaded from http://www.protectedplanet.net/search?marine=1 in March 2015}
}
\description{
Title
}

