% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest.R
\name{harvest}
\alias{harvest}
\title{Title}
\usage{
harvest(fish, quota, ages = min_age_catch:maxage, distance, fish_licenses,
  mpaSQ, mpa2020, writefish = TRUE, writefishcatch = TRUE, writekg = TRUE,
  results_folder)
}
\arguments{
\item{writekg}{}
}
\description{
Title
}

