% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initcm.R
\name{initcm}
\alias{initcm}
\title{Initiate the connectivity matrices}
\usage{
initcm(domain, e_fold, cell_size)
}
\arguments{
\item{domain}{the gridded model domain (SpatialPolygonsDataFrame) as generated by inigrid}

\item{e_fold}{The e-folding scale for dispersal in m (the distance at which there will be fewer larvae/adults dispersed from a central point by a factor of e)}

\item{cell_size}{cell size of the grid used in initgrid}
}
\description{
Initiate the connectivity matrices
}
\examples{
initcm(domain=BESTMPA_domain,e_fold=75000,cell_size=20000)
}

