% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initpop.R
\name{initpop}
\alias{initpop}
\title{Initialize population matrix}
\usage{
initpop(initial_abun = 250 * 10^6, cells = 5000, maxage = 50,
  rate = 0.9)
}
\arguments{
\item{rate}{}
}
\description{
Initialize population matrix
}
\examples{
fish <- initpop(initial_abun=250*10^6,cells=5000,maxage=50,rate=0.9)
}

