% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressBar.R
\name{progressBar}
\alias{progressBar}
\title{Title}
\usage{
progressBar(title = "R progress bar", label = "", min = 0, max = 1,
  initial = 0, width = 300)
}
\arguments{
\item{width}{}
}
\description{
Title
}

