% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduction.R
\name{reproduction}
\alias{reproduction}
\title{Title}
\usage{
reproduction(fish, fecundity, age_mat_steepness, age_mat_sigmoid, l_to_w_int,
  l_to_w_power, Linf_mean, Linf_SD, k_mean, k_SD, t0, breeding_near, p)
}
\arguments{
\item{p}{}
}
\description{
Title
}

