% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundprob.R
\name{roundprob}
\alias{roundprob}
\title{Rounds based on probability}
\usage{
roundprob(x)
}
\arguments{
\item{x}{a vector of numbers to be rounded}
}
\description{
This will round `x` probabilistically. For example, `roundprob(1.5)` will return `1` in roughly half of the cases and will return `2` in the other cases.
}
\examples{
table(roundprob(rep(1.5,100000)))
}

