% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setProgressBar.R
\name{setProgressBar}
\alias{setProgressBar}
\title{Set progress bar}
\usage{
setProgressBar(pb, value, title = NULL, label = NULL)
}
\arguments{
\item{pb}{an object of class \code{"winProgressBar"}.}

\item{title}{character strings, giving the window title and the
    label on the dialog box respectively.}
}
\description{
OS agnostic wrapper function for utils::setWinProgressBar and tcltk::setTkProgressBar
}

