% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addscenario.R
\name{addscenario}
\alias{addscenario}
\title{Title}
\usage{
addscenario(domain, included = NA, priority = NA, excluded = NA,
  MPA_coverage = 0.1, replicates = 25, dist = NA, name = "MPA_SQ",
  cell_size, cells)
}
\arguments{
\item{domain}{the gridded model domain (SpatialPolygonsDataFrame) as generated by inigrid}

\item{included}{}

\item{priority}{a logical vector indicating which grid cells are to be prioritized when assigning MPA protection. The algorithm will select all}
}
\description{
Title
}

