% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addscenario.R
\name{addscenario}
\alias{addscenario}
\title{Title}
\usage{
addscenario(domain, included = NA, priority = NA, excluded = NA,
  MPA_coverage = 0.1, replicates = 10, dist = NA, name = "MPA_SQ",
  cell_size, cells)
}
\arguments{
\item{included}{}
}
\description{
Title
}

