% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addscenario.R
\name{addscenario}
\alias{addscenario}
\title{Add a protection scenario}
\usage{
addscenario(domain, included = NA, priority = NA, excluded = NA,
  MPA_coverage = 0.1, replicates = 25, dist = NA, name = "MPA_SQ",
  cell_size, cells = 1)
}
\arguments{
\item{domain}{the gridded model domain (SpatialPolygonsDataFrame) as generated by inigrid}

\item{included}{A logical vector indicating which grid cells are to be included in MPA assignment. This accomodates the pre-existing MPAs.}

\item{priority}{A logical vector indicating which grid cells are to be prioritized when assigning MPA protection. The algorithm will select all priority cells regardless of distance/spacing before selecting non-priority cells.}

\item{excluded}{A logical vector indicating which grid cells are to be excluded from MPA assignment}

\item{MPA_coverage}{Target protection level as a proportion (e.g. 0.2 is 20\% protection)}

\item{replicates}{A vector of replicates for the main body of the model}

\item{dist}{The fixed distance for setting MPA distance in m in fixed distance scenarios.}

\item{name}{The name of the scenario to be added. While any character string will be accepted, using the format "MPA_XX" where XX is any 2 letters which indicates treatment type is suggested to minimize conflict with other functions in the BESTMPA package.}

\item{cell_size}{cell size of the grid used in initgrid}

\item{cells}{minimum number of cells that can be considered as an MPA. defaults to 1}
}
\value{
SpatialPolygonsDataFrame
}
\description{
Adds protection scenarios to the SpatialPolygonsDataFrame grid
}
\examples{
proj  <- "+proj=utm +zone=20 +datum=NAD83 +units=m +no_defs +ellps=GRS80 +towgs84=0,0,0"
p <- initgrid(EEZ=EEZ,cell_size=20000,proj=proj,areas=c(Habitats=Habitats,Breeding=Breeding))
# set targeted scenario
p <- addscenario(domain=p,included=oldMPA,priority=p$Habitats,excluded=p$Breeding,MPA_coverage=0.1,replicates=1:3,dist=75000,name="MPA_TR",cell_size=20000,cells=3)
}

