% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age2length.R
\name{age2length}
\alias{age2length}
\title{Convert fish age to length}
\usage{
age2length(t = 0:50, Linf_mean = 112.03, Linf_SD = 10.46/1.96,
  k_mean = 0.13, k_SD = 0.021/196, t0 = 0.18)
}
\arguments{
\item{t}{is the age of the fish}

\item{t0}{}
}
\description{
Von Bertalanffy growth model parameters (Knickle and Rose 2013)
Lt = Linf * {1-exp(-k*(t-t0))}, where Lt is length (cm) at age t (year), Linf is the asymptotic length (cm), k is the VB growth coefficient (1/year), and t0 is the x intercept (year). Linf = 112.03 (95% CI = 10.46). k = 0.13 (95% CI = 0.021). t0 = 0.18).
}

