% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal.R
\name{dispersal}
\alias{dispersal}
\title{Adult or larval dispersal}
\usage{
dispersal(fish, cm, ages = 5:50)
}
\arguments{
\item{fish}{a matrix of fish numbers by site (rows) and by age (columns). Can easily be generated by \code{initpop}}

\item{cm}{a connectivity matrix containing the probability of dispersing from each starting location (rows) to each end location (columns). Can easily be generated by \code{initcm}}

\item{ages}{age of fish to be dispersed}
}
\description{
Disperses individuals of specified \code{ages} according to a user defined connectivity matrix.
}
\examples{

cells=5000

# homogeneous dispersal
cm <- matrix(1/cells,nrow=cells,ncol=cells)
fish <- initpop(initial_abun=250*10^6,cells=cells,maxage=50,rate=0.7)
fish <- dispersal(fish,cm,ages=5:50)

# exponential dispersal kernal
cm <- initcm(domain=BESTMPA_domain,e_fold=75000,cell_size=20000)
fish <- initpop(initial_abun=250*10^6,cells=length(BESTMPA_domain),maxage=50,rate=0.7)
fish <- dispersal(fish,cm,ages=5:50)
}

