% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish2eggs.R
\name{fish2eggs}
\alias{fish2eggs}
\title{Generate eggs from fish}
\usage{
fish2eggs(fish, fecundity = 0.5 * 10^6, age_mat_steepness = 2.5,
  age_mat_sigmoid = 4, l_to_w_int = 1.1e-05, l_to_w_power = 2.91,
  Linf_mean = 112.03, Linf_SD = 10.46/1.96, k_mean = 0.13,
  k_SD = 0.021/196, t0 = 0.18)
}
\arguments{
\item{fish}{a matrix of fish numbers by site (rows) and by age (columns). Can easily be generated by \code{initpop}}

\item{fecundity}{Size dependent fecundity (eggs per kg of female)}

\item{age_mat_steepness}{Steepness of the logistic curve for age at sexual maturity. By default, fish begin maturing at 2 y, 50\% at 4 y and all are mature at 6y}

\item{age_mat_sigmoid}{Sigmoid of the logistic curve for age at sexual maturity. By default, fish begin maturing at 2 y, 50\% at 4 y and all are mature at 6y}

\item{l_to_w_int}{The intercept in the Length-weight relationship}

\item{l_to_w_power}{The power in the Length-weight relationship}

\item{Linf_mean}{Von Bertalanffy growth model parameter - mean asymptotic length (cm)}

\item{Linf_SD}{Von Bertalanffy growth model parameter - standard deviation asymptotic length (cm)}

\item{k_mean}{Von Bertalanffy growth model parameters - mean growth coefficient (1/year)}

\item{k_SD}{Von Bertalanffy growth model parameters - standard deviation growth coefficient (1/year)}

\item{t0}{Von Bertalanffy growth model parameters - x intercept (year)}
}
\description{
Generates eggs from fish, eggs remain in parents' cell Number of eggs produced is based on the age to length to weight relationships, the age at maturity, and fecundity levels.
}
\examples{
# create a fish matrix
fish <- initpop(initial_abun=250*10^6,cells=length(BESTMPA_domain),maxage=50,rate=0.7)
# create eggs
eggs <- fish2eggs(fish,fecundity=0.5*10^6,age_mat_steepness=2.5,age_mat_sigmoid=4,l_to_w_int=0.000011,l_to_w_power=2.91,Linf_mean=112.03,Linf_SD=10.46/1.96,k_mean=0.13,k_SD=0.021/196,t0=0.18)
}

