% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest.R
\name{harvest}
\alias{harvest}
\title{Title}
\usage{
harvest(fish, quota, ages = min_age_catch:maxage, distance, fish_licenses,
  mpabefore, mpaafter, writefish = TRUE, writefishcatch = TRUE,
  writekg = TRUE, results_folder, y, time, suffix = NA)
}
\arguments{
\item{fish}{a matrix of fish numbers by site (rows) and by age (columns). Can easily be generated by \code{initpop}}

\item{quota}{fishable quota, can be estimated using \code{link{estimatequota}}}

\item{ages}{age of catchable fish}

\item{distance}{matrix of distance from each cell to each \code{fish_communities}}

\item{fish_licenses}{number of licenses per region in \code{fish_communities}}

\item{mpabefore}{Logical vector, TRUE if closed to fishing before \code{time}}

\item{mpaafter}{Logical vector, TRUE if closed to fishing during \code{time}}

\item{writefish}{If TRUE, will write the \code{fish} remaining for each cell/age to the \code{results_folder}}

\item{writefishcatch}{If TRUE, will write the \code{fish} caught for each cell/age to the \code{results_folder}}

\item{writekg}{(logical) If TRUE, will write the number of kg of fish caught for each cell/community to the \code{results_folder}}

\item{results_folder}{Path to the directory where results are stored}

\item{y}{current year}

\item{time}{years during which model results are recorded (i.e. not during spin-up time)}
}
\description{
Title
}
\examples{
#' # create a fish matrix
fish <- initpop(initial_abun=250*10^6,cells=length(BESTMPA_domain),maxage=50,rate=0.7)
# calculate distances from shore
distance <- gDistance(spTransform(fish_communities,proj),BESTMPA_domain,byid = T)
# number of licenses per region in fish_communities
fish_licenses <- c(866, 4714, 3002, 879, 963)
# define MPAs
mpa <- sample(c(TRUE,FALSE),length(BESTMPA_domain),replace=TRUE)
nompa <- rep(FALSE,length(BESTMPA_domain))
# quota
quota <- estimatequota(fish,maxage=20,y=2001,tot_time=2001:2071,FMSY=0.28,FMSY_buffer=0.667)
# go fishing!
fish <- harvest(fish,quota,ages=4:20,distance,fish_licenses,mpabefore=nompa,mpaafter=mpa,writefish=FALSE,writefishcatch=FALSE,writekg=TRUE,results_folder='',y=2001,time=time)
}

