% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{housekeeping}
\alias{housekeeping}
\title{Title}
\usage{
housekeeping(results_folder = "output", env = FALSE, fig = FALSE,
  delete = FALSE, memorylimit = memory.limit())
}
\arguments{
\item{memorylimit}{}
}
\description{
Title
}

