% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduction.R
\name{reproduction}
\alias{reproduction}
\title{Fish reproduction}
\usage{
reproduction(fish, fecundity, age_mat_steepness, age_mat_sigmoid, l_to_w_int,
  l_to_w_power, Linf_mean, Linf_SD, k_mean, k_SD, t0, breeding_near, domain)
}
\arguments{
\item{fish}{a matrix of fish numbers by site (rows) and by age (columns). Can easily be generated by \code{initpop}}

\item{l_to_w_int}{The intercept in the Length-weight relationship}

\item{l_to_w_power}{The power in the Length-weight relationship}

\item{Linf_mean}{Von Bertalanffy growth model parameter - mean asymptotic length (cm)}

\item{Linf_SD}{Von Bertalanffy growth model parameter - standard deviation asymptotic length (cm)}

\item{k_mean}{Von Bertalanffy growth model parameters - mean growth coefficient (1/year)}

\item{k_SD}{Von Bertalanffy growth model parameters - standard deviation growth coefficient (1/year)}

\item{t0}{Von Bertalanffy growth model parameters - x intercept (year)}

\item{breeding_near}{vector of nearest Breeding}

\item{domain}{the gridded model domain (SpatialPolygonsDataFrame) as generated by inigrid}
}
\description{
Adult fish will release eggs in the nearest breeding ground. Number of eggs produced is based on the age to length to weight relationships, the age at maturity, and fecundity levels.
}
\examples{
# create a fish matrix
fish <- initpop(initial_abun=250*10^6,cells=length(BESTMPA_domain),maxage=50,rate=0.7)
# identify nearby breeding grounds
breeding_near <- apply(gDistance(spTransform(Breeding,proj),p,byid = T),1,which.min)
# reproduction
recruits <- reproduction(fish,fecundity=5e+05,age_mat_steepness=2.5,age_mat_sigmoid=4,l_to_w_int=1.1e-05,l_to_w_power=2.91,Linf_mean=112.03,Linf_SD=5.336735,k_mean=0.13,k_SD=0.01071429,t0=0.18,breeding_near=breeding_near,domain=BESTMPA_domain)
}

